Refactor: Remove functions from  async waterfall
Currently, each function for the network request is wrapped within the waterfall method from the async library.
Each function produces a callback that is in turned passed to the next function.
However, to isolate the network request, I wanted to "unwrap" the functions from this waterfall.

The primary purpose of this refactor was to follow the single responsibility purpose more closely and isolate
each function's purpose within the larger set of functions.

I was also able to remove use of follow-redirects and instead rely on fetch to get the URL that a TikTok URL
redirects to. I restrategized how I determined a TikTok video's information is available. Instead of checking
to see if a particular field (i.e. title) is filled out, I instead check to see if there is a status_msg
key with a value of 'Something went wrong.'