Feat: Refactor/recreate adding videos to database
Change URL added to video when calling getVideoDataObject()

Previously, I used the URL string from the user's own data to add to the video object
that would later be inserted into the database. However, this URL DOES NOT meet the
constraints that I placed on URLs inserted into the database. The URL format is slightly
different from what is allowed. The URL that I actually want to add is the redirected
URL based off of the URL in the user's own data.

The format TikTok uses for its URLs can be formulated as: the author's url + /video/ +
video ID number. This is the correct format for being added into the database.
I recreated this format and added it as the url property to the video object.

I also realized that I was not adding in information about the date a user had bookmarked
a video, so I incorporated that into the processVideos() function as well.

After correcting this issue, I created a series of functions that I captured within an
addVideo() function that ask true/false questions, including a function to check a user's videos for,
and a function to check all videos for duplicates. I also house a function to add a video to
all videos and a function to add a video just to users videos. I call
this function within the processVideos() function to add videos to the database when/how
applicable.