Test: Test addVideo() function
After refactoring the addVideo() function, I decided it would make the most sense
to only test the addVideo() function when testing the database.
Testing the functions that add videos to the database (i.e. addVideoToAllVideos() and addVideoToUsersVideos())
would mostly just mean trying to test Postgres. Likewise, although I could test the functions that check if
a video exists in the database, those tests may not be particularly valuable or confidence-worthy because they only
return true/false based on the data that I give them.

I am more interested in testing the addVideo() function because I want to test the logic of the function
to confirm that different inputs hit the correct paths.

To do this, I mocked out the functions that can be called inside of the addVideo() function (i.e.
addVideoToAllVideos() and addVideoToUsersVideos()).

I began by testing a video that did not exist in the user's videos table and did exist in the general videos table either.
I expect both addVideoToAllVideos() and addVideoToUsersVideos() to be called.

I tested the following situations as well:
- A video exists in all videos but not a users videos table
- A video exists in both the users videos and all videos table

When a video exists in both the users videos table and the videos table, it should also return a message.
I tested for that message return as well.

To avoid calls from previous tests throwing off other tests, I added a beforeEach() function to clear the mocked function calls.
I also removed console messages from my database tests to reduce noise from tests that may be confused with actual application
messages.