Feat: Update array of error messages
I previously created an array of error messages to use in my application, and I am now
adding additional messages to replace existing messages when adding multiple videos to
the databasuse in my application.

I tried to make error messages more reusable while still conveying at least some relevant
information to the user. I also did not want to expose too much information to the user,
in the event that the user is malicious.

The invalidFile key can be used when a user uploads a file with the wrong name or a file
that contains characters that are invalid. The emptyFile key can be used when a user
uploads an empty file. To contrast, the database key can be used in about four different
scenarios: when there is an error adding a video to a user's videos list, when there is an
error adding a video to the general videos list, when there is an error checking for
duplicates in the users videos list, and when there is an error checking for duplicates in
the general videos list.