Feat: Convert array into array of error objects
I decided to create an array of error message objects. This allowed me to make
error messages more specific but still categorized. I differentiated between files
that are invalid because they are empty, files that are invalid because they have
invalid characters, and files that are invalid because of their extension/size/name.
I also separated database errors into errors reading the database and errors writing
to the database.

My hope is that I can use this array to create error messages that may be more useful
to the user on the client-side of the application.