Style: Fix linter warnings
There were over 900 linter warnings in my videoController file.
This made ESLint significantly less useful because linter warnings from new code
I was adding were difficult to find.

I corrected essentially all auto-fixable problems and temporarily disabled some
linter warnings so as to be able to focus on the new warnings that could arise from
newer code.

Additionally, I left one linter warning in place (a warning about a return at the end
of an arrow function) because it may have an implication in how I structure some of
my error handling functions in the future. This will, alongside the disabled warnings,
will be left for either future refactoring or to be handled when further features for
error handling are implemented.