Feat: Create wrapper for logging to Rollbar
I created a wrapper function called logErrors that would accept an error level, information about an error, and parameters
(if need be). The function logs this information to Rollbar.

I called the logErrors() function after a user logs in to the app. I called the function
at the error level, sent a test error, and sent two test parameters.

I confirmed that the error appeared on the Rollbar website.

FSA2021-153