Fix: Update safe guard for duplicate videos
While I was testing the behavior of the functions for adding multiple videos when
you remove the try/catch block from the checkUsersVideosForDuplicates() function, I noticed
that returning undefined would mean that when checkUsersVideosForDuplicates() is stored
as a variable (isUsersVideosDuplicate), a value of undefined would mean that isUsersVideosDuplicate
is false, or !isUsersVideosDuplicate.

When isUsersVideosDuplicate is false, a function to add the video to the users videos will be called,
at the very least. I asked myself if I wanted this to be the default behavior of the
checkUsersVideosForDuplicates() function when there are errors. If there are errors, I do not
want any duplicates to be added. I would prefer to NOT add new videos rather than duplicate existing
videos.

To satisfy this, I changed the return statement from undefined to true. Just to be safe,
we will assume that a video is a duplicate if we cannot prove otherwise.

checkAllVideosForDuplicates() has a similar issue to checkUsersVideosForDuplicates() with the
isAllVideosDuplicate variable, so I also updated its return value to true instead of undefined.