Refactor: Replace link with button
I previously created a button for returning a user to the list of videos. However, this
link doesn't actually go anywhere. It operates by using a bit of JavaScript to change the
CSS for the page where videos are visible and the user account information is not.

Because this is not an anchor and is a UI update, I changed this from a link to a button.

FSA2021-186