Refactor: Create header for user account
I decided it would make the most sense for the h2 element to appear first
on the user account page, semantically. However, visually, I would like
the button returning the user to their videos to appear first. I am
accomplishing this by using a flex display on the header container and
changing the order of its child components.

FSA2021-186