Refactor: Use BEM in header names
I implemented BEM here to organize the content of the header better.
I added an additional class to the header tag itself of account-details.
I used this account-details class to create elements for the h2 element
and the paragraph element.
I am using the user-account class to create an element for the button.

FSA2021-186