Refactor: Update class names
For the buttons, I am using the user-account block to create a form-toggle element.
This will enable me to apply similar styles to all of the buttons that can be used
to toggle different user account options.

Additionally, the button also has a class that concerns only its particular action
with an element of button. Likewise, the material icons arrow has a block unique
to its section and an element of arrow.

FSA2021-186