Refactor: Update form class name
To adhere more to BEM, I renamed the forms to use the block__element format.

However, there is still a form group class here that is used on multiple forms
throughout the website. I would like to use BEM and ITCSS for these reused
form classes, but that would be part of a broader restructuring, so I think that
work would have to be done in a separate PR.

FSA2021-186