Refactor: Move button style to mixin
I created a mixins file inside of a new tools folder and imported that into
my stylesheet. I want to use the same styling to make all of the buttons on the
page appear like links, so to maximize reusability, I created a mixin based
off of the styles I created for the user account form toggles to accomplish this.

FSA2021-186