Style: Remove linter warnings from userController
Because ESLint was added to this project only recently, there were a lot
of warnings/errors in the linter that needed to be remedied before the linter
could be of any use. I was able to remove most of these linter warnings
more or less automatically, but I temporarily disabling consistent return
(i.e. currently there are no return statements on several arrow functions)
until I come around to do a more extensive refactor.

FSA2021-182