Feat: Integrate flash messages
I realized that my flash messages were not working was because I was not
passing in the word 'error' as a string using the req.flash() method.
After doing this when rendering my login page and adding a failureFlash
property with a value of true when posting login requests, I can now see
flash messages.

There are several flash messages interspersed throughout the userController.js file.
I took the majority of these flash messages and added their message as a key-value pairs
in an object dedicated to error messages.
This way, I can increase reusability and reduce errors caused by misspellings in strings.

I previously anticipated using flash messages to convey non-error feedback to the user.
However, after these flash messages come from an external package and require a specific
format in order to use, i.e. req.flash('key', 'flash message value') and
res.send({ messages: req.flash('key') }). Each time I wanted to display a flash message
at a particular endpoint, I would have to set it up this way.

Because this would mean increasing my exposure to an external API and making it more
difficult to use reusable variables, I asked myself if it made sense to use flash messages
in these situations. Unlike my login system, which uses passport and can only display
flash messages by using an external package, I do not have this limitation in other areas
of my application.

Therefore, when flashing messages to the user that do not intersect with my passport.js
login system, I will pass the values of these messages directly to the messages object.

Right now, the user feedback that is included in this PR comes from a post
request that is made on a page dedicated to user account details. However,
its get request is the same as the index page's (i.e. navigating to this page
is technically a UI change, not going to a new link). Because of this, if
I want to render the same page (i.e. the index view  page), I will get an error
because the URL path will be that of the user account details post request,
not the index page.

The URL where the post request is located cannot access all of the resources
that the index page typically gets (i.e. when loaded through the get
request for the index page). When I try to render the index page, I receive
an error.

To avoid this, I am rendering the error page instead and loading the message
information into this error page. The error page did not previously accept
the messages property, so I added it in. If there are messages, they will load.
If there are errors, the errors will load. If messages do not exist, they
will not load. If errors do not exist, those will not load.

In a future PR, I will try to separate the user account details page from
the index page entirely as two separate routes/get requests. Then, I will
render the message inside of the user account details page, which should
be located along the same path as the user account details page post request.
I will also be able to use th error view page only for errors, in this
future PR.

FSA2021-182