Feat: Create function for email matches
Before I delete a user's account, I want to make sure that
the email address they entered matches the email address that
they are registered with. To do this, I created an async
function that will query my database to confirm that their
email entry matches the email for their user ID.
The function will return a Boolean value of true if the
emails match (i.e. if there is a result for their user ID
with the email address they entered). Otherwise, it will return
false.

FSA2021-187