Fix: Update password
-- Remove async.waterfall() wrapping
-- Rename hashNewPassword() to hashPassword() and make it more modular for use
in other areas of the app
-- Rename existingUserPassword() to getUserPasswordFromUserId() for clarity
and reusability
-- Rename and refactor existingPasswordMatch() to checkPasswordMatch() for
clarity. Give it parameters instead of directly feeding Express objects to it.
-- Update error handling/logging throughout all functions
-- Refactor and rename updatePassword() to handleUpdatePasswordRequest() because
this function can update a user's password but it can also indicate a user error
with a mismatched password. This function also accepts parameters and calls
the functions previously created in this commit.
-- Add backend form validation to handle user-generated errors.
-- Abstract try/catch down to only be in our final call to the
handleUpdatePasswordRequest() function

The cause of this error was not waiting for a promise to resolve before trying
to use its value.

FSA2021-189